refactor: rename classes
