+++ coordination/src/main/java/tech/ydb/coordination/CoordinationSession.java
@@ -1,146 +1,156 @@
 package tech.ydb.coordination;
 
+import java.time.Duration;
 import java.util.concurrent.CompletableFuture;
+import java.util.function.Consumer;
 
+import tech.ydb.coordination.settings.DescribeSemaphoreChanged;
 import tech.ydb.coordination.settings.SemaphoreDescription;
+import tech.ydb.core.Result;
 import tech.ydb.core.Status;
-import tech.ydb.proto.coordination.SessionRequest;
 
-/**
- * @author Kirill Kurdyukov
- */
-public interface CoordinationSession {
+public interface CoordinationSession extends AutoCloseable {
 
-    long getSessionId();
-
-    byte[] getProtectionKey();
-
-//    CompletableFuture<Status> start(CoordinationSession.Observer observer, String nodePath);
+    @Override
+    void close();
 
     /**
-     * First message used to start/restore a session
-     *
-     * @param sessionStart session start of proto body
+     * {@link CoordinationSession#createSemaphore(String, long, byte[])}
      */
-    @Deprecated
-    void sendStartSession(SessionRequest.SessionStart sessionStart);
+    default CompletableFuture<Status> createSemaphore(String semaphoreName, long limit) {
+        return createSemaphore(semaphoreName, limit, null);
+    }
 
     /**
-     * Used for checking liveness of the connection
+     * Used to create a new semaphore
      *
-     * @param pingPong ping pong of proto body
+     * @param semaphoreName Name of the semaphore to create
+     * @param limit         Number of tokens that may be acquired by sessions
+     * @param data          User-defined data that is attached to the semaphore
      */
-    @Deprecated
-    void sendPingPong(SessionRequest.PingPong pingPong);
+    CompletableFuture<Status> createSemaphore(String semaphoreName, long limit, byte[] data);
 
-    void sendAcquireSemaphore(String semaphoreName, long count, long timeoutMillis, boolean ephemeral);
-    void sendAcquireSemaphore(String semaphoreName, long count, long timeoutMillis, boolean ephemeral, long requestId);
+
+    default CompletableFuture<Result<CoordinationSemaphore>> acquireSemaphore(
+            String semaphoreName, long count, boolean ephemeral, Duration timeout) {
+        return acquireSemaphore(semaphoreName, count, ephemeral, timeout, null);
+    }
 
     /**
-     * Used to acquire a semaphore
+     * {@link CoordinationSession#acquireSemaphore(String, long, boolean, Duration, byte[])}
+     */
+    default CompletableFuture<Result<CoordinationSemaphore>> acquireSemaphore(
+            String semaphoreName, long count, Duration timeout) {
+        return acquireSemaphore(semaphoreName, count, false, timeout, null);
+    }
+
+    /**
+     * Used to acquire an ephemeral semaphore.
+     * Ephemeral semaphores are created with the first acquire operation and automatically deleted with
+     * the last release operation.
      * <p>
      * WARNING: a single session cannot acquire the same semaphore multiple times
      * <p>
      * Later requests override previous operations with the same semaphore,
      * e.g. to reduce acquired count, change timeout or attached data
+     *
      * @param semaphoreName Name of the semaphore to acquire
-     * @param count Number of tokens to acquire on the semaphore
-     * @param timeoutMillis Timeout in milliseconds after which operation will fail
+     * @param count         Number of tokens to acquire on the semaphore
+     * @param timeout       Duration after which operation will fail
      *                      if it's still waiting in the waiters queue
-     * @param ephemeral Ephemeral semaphores are created with the first acquire operation
-     *                      and automatically deleted with the last release operation
-     * @param data User-defined binary data that may be attached to the operation
-     * @param requestId Client-defined request id, echoed in the response
+     * @param data          User-defined binary data that may be attached to the operation
      */
-    void sendAcquireSemaphore(String semaphoreName, long count, long timeoutMillis, boolean ephemeral, byte[] data,
-                              long requestId);
+    CompletableFuture<Result<CoordinationSemaphore>> acquireSemaphore(String semaphoreName, long count,
+                                                                      boolean ephemeral,
+                                                                      Duration timeout, byte[] data);
 
-    void sendReleaseSemaphore(String semaphoreName);
+    CompletableFuture<Status> updateSemaphore(String semaphoreName, byte[] data);
 
-    /**
-     * Used to release a semaphore
-     * <p>
-     * WARNING: a single session cannot release the same semaphore multiple times
-     * </p>
-     * The release operation will either remove current session from waiters
-     * queue or release an already owned semaphore.
-     * @param semaphoreName Name of the semaphore to release
-     * @param requestId Client-defined request id, echoed in the response
-     */
-    void sendReleaseSemaphore(String semaphoreName, long requestId);
+    CompletableFuture<Result<SemaphoreDescription>> describeSemaphore(String semaphoreName, DescribeMode describeMode,
+                                                                      WatchMode watchMode,
+                                                                      Consumer<DescribeSemaphoreChanged> updateWatcher);
 
-    void sendDescribeSemaphore(String semaphoreName);
+    CompletableFuture<Result<SemaphoreDescription>> describeSemaphore(String semaphoreName, DescribeMode mode);
 
-    /**
-     * Used to describe semaphores and watch them for changes
-     * <p>
-     * WARNING: a describe operation will cancel previous watches on the same semaphore
-     * </p>
-     * @param semaphoreName Name of the semaphore to describe
-     * @param includeOwners Response will include owners list if true
-     * @param includeWaiters Response will include waiters list if true
-     * @param watchData Watch for changes in semaphore data
-     * @param watchOwners Watch for changes in semaphore owners (including owners' data)
-     */
-    void sendDescribeSemaphore(String semaphoreName, boolean includeOwners, boolean includeWaiters, boolean watchData,
-                               boolean watchOwners);
+    CompletableFuture<Status> deleteSemaphore(String semaphoreName, boolean force);
 
-    void sendCreateSemaphore(String semaphoreName, long limit);
+    long getId();
 
-    /**
-     * Used to create a new semaphore
-     * @param semaphoreName Name of the semaphore to create
-     * @param limit Number of tokens that may be acquired by sessions
-     * @param data User-defined data that is attached to the semaphore
-     * @param requestId Client-defined request id, echoed in the response
-     */
-    void sendCreateSemaphore(String semaphoreName, long limit, byte[] data, long requestId);
+    enum DescribeMode {
+        /**
+         * Describe only semaphore's data (name, user-defined data and others)
+         */
+        DATA_ONLY(false, false),
+        /**
+         * Include owners list to describe result
+         */
+        WITH_OWNERS(true, false),
+        /**
+         * Include waiters list to describe result
+         */
+        WITH_WAITERS(false, true),
+        /**
+         * Include waiters and owners lists to describe result
+         */
+        WITH_OWNERS_AND_WAITERS(true, true);
 
-    void sendUpdateSemaphore(String semaphoreName, byte[] data);
+        private final boolean includeOwners;
+        private final boolean includeWaiters;
 
-    /**
-     * Used to change semaphore data
-     * @param semaphoreName Name of the semaphore to update
-     * @param data User-defined data that is attached to the semaphore
-     * @param requestId Client-defined request id, echoed in the response
-     */
-    void sendUpdateSemaphore(String semaphoreName, byte[] data, long requestId);
-
-    void sendDeleteSemaphore(String semaphoreName, boolean force);
-
-    /**
-     * Used to delete an existing semaphore
-     * @param semaphoreName Name of the semaphore to delete
-     * @param force Will delete semaphore even if currently acquired by sessions
-     * @param requestId Client-defined request id, echoed in the response
-     */
-    void sendDeleteSemaphore(String semaphoreName, boolean force, long requestId);
-
-    CompletableFuture<Status> getLifetimeFuture();
-
-    void stop();
-
-    interface Observer {
-
-        default void onAcquireSemaphoreResult(boolean acquired, Status status, int requestId) {
+        DescribeMode(boolean includeOwners, boolean includeWaiters) {
+            this.includeOwners = includeOwners;
+            this.includeWaiters = includeWaiters;
         }
 
-        default void onDescribeSemaphoreResult(SemaphoreDescription description, Status status, int requestId) {
+        public boolean includeOwners() {
+            return includeOwners;
         }
 
-        default void onDescribeSemaphoreChanged(boolean dataChanged, boolean ownersChanged, int requestId) {
+        public boolean includeWaiters() {
+            return includeWaiters;
+        }
+    }
+
+    enum WatchMode {
+        /**
+         * Watch for changes in semaphore data
+         */
+        WATCH_DATA(true, false),
+        /**
+         * Watch for changes in semaphore owners
+         */
+        WATCH_OWNERS(false, true),
+        /**
+         * Watch for changes in semaphore data or owners
+         */
+        WATCH_DATA_AND_OWNERS(true, true);
+
+        private final boolean watchData;
+        private final boolean watchOwners;
+
+        WatchMode(boolean watchData, boolean watchOwners) {
+            this.watchData = watchData;
+            this.watchOwners = watchOwners;
         }
 
-        default void onDeleteSemaphoreResult(Status status, int requestId) {
+        public boolean watchData() {
+            return watchData;
         }
 
-        default void onCreateSemaphoreResult(Status status, int requestId) {
+        public boolean watchOwners() {
+            return watchOwners;
         }
+    }
 
-        default void onReleaseSemaphoreResult(boolean released, Status status, int requestId) {
-        }
-
-        default void onUpdateSemaphoreResult(Status status, int requestId) {
-        }
+    interface CoordinationSemaphore {
+        /**
+         * Used to release a semaphore
+         * <p>
+         * WARNING: a single session cannot release the same semaphore multiple times
+         * </p>
+         * The release operation will either remove current session from waiters
+         * queue or release an already owned semaphore.
+         */
+        CompletableFuture<Result<Boolean>> release();
     }
 }
